X/requirements.txt
feedparser
tweepy
anthropic
google-generativeai
python-dotenv
requests
